Fixed Issue #1 by Moving Template and Static folders to inside Blog package.
