+++ blog/templates/auth/register.html
